# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Raudah Travels & Tours API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/profile:
    get:
      operationId: getProfile
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Profile not found
    put:
      operationId: updateProfile
      tags: [auth]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /auth/profile/sync:
    post:
      operationId: syncProfile
      tags: [auth]
      summary: Sync Clerk user to DB profile (upsert on login)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncProfileBody"
      responses:
        "200":
          description: Profile synced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  # --- Packages ---
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings (filtered by user role)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, cancelled, completed]
        - name: userId
          in: query
          schema:
            type: string
        - name: packageId
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: number
        - name: offset
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingListResponse"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Created booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    put:
      operationId: updateBooking
      tags: [bookings]
      summary: Update booking status or details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingBody"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  # --- Payments ---
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationListResponse"
    post:
      operationId: createNotification
      tags: [notifications]
      summary: Create notification (admin/system)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNotificationBody"
      responses:
        "201":
          description: Notification created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        clerkUserId:
          type: string
        email:
          type: string
        fullName:
          type: string
        phone:
          type: string
        avatarUrl:
          type: string
          nullable: true
        role:
          type: string
          enum: [user, agent, staff, admin, super_admin]
        ninNumber:
          type: string
          nullable: true
        passportNumber:
          type: string
          nullable: true
        dateOfBirth:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, clerkUserId, email, fullName, role, createdAt]
    UpdateProfileBody:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        avatarUrl:
          type: string
        ninNumber:
          type: string
        passportNumber:
          type: string
        dateOfBirth:
          type: string
        address:
          type: string
    SyncProfileBody:
      type: object
      properties:
        clerkUserId:
          type: string
        email:
          type: string
        fullName:
          type: string
        avatarUrl:
          type: string
      required: [clerkUserId, email, fullName]
    TravelPackage:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [hajj, umrah]
        category:
          type: string
          enum: [premium, standard, budget]
        season:
          type: string
          nullable: true
        description:
          type: string
        price:
          type: number
        agentDiscount:
          type: number
        depositAmount:
          type: number
        durationDays:
          type: number
        departureDate:
          type: string
        returnDate:
          type: string
        departureCities:
          type: array
          items:
            type: string
        airlines:
          type: array
          items:
            type: string
        maxCapacity:
          type: number
        currentBookings:
          type: number
        inclusions:
          type: array
          items:
            type: string
        imageUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        starRating:
          type: number
        status:
          type: string
          enum: [active, draft, archived]
        countdownEnabled:
          type: boolean
        countdownExpiry:
          type: string
          nullable: true
        countdownAction:
          type: string
          enum: [disable, both]
        isRegistrationClosed:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - type
        - category
        - description
        - price
        - agentDiscount
        - depositAmount
        - durationDays
        - departureDate
        - returnDate
        - departureCities
        - airlines
        - maxCapacity
        - currentBookings
        - inclusions
        - isActive
        - isFeatured
        - starRating
        - status
        - countdownEnabled
        - countdownAction
        - isRegistrationClosed
        - createdAt
...
    Booking:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        packageId:
          type: string
        agentId:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        totalPrice:
          type: number
        amountPaid:
          type: number
        pilgrimCount:
          type: number
        pilgrimDetails:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
        package:
          $ref: "#/components/schemas/TravelPackage"
          nullable: true
        user:
          $ref: "#/components/schemas/UserProfile"
          nullable: true
      required: [id, userId, packageId, status, totalPrice, amountPaid, pilgrimCount, createdAt, updatedAt]
    CreateBookingBody:
      type: object
      properties:
        packageId:
          type: string
        pilgrimCount:
          type: number
        pilgrimDetails:
          type: string
        notes:
          type: string
        agentId:
          type: string
        fullName:
          type: string
          nullable: true
        passportNumber:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
      required: [packageId, pilgrimCount]
    UpdateBookingBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        notes:
          type: string
        amountPaid:
          type: number
    BookingListResponse:
      type: object
      properties:
        bookings:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
        total:
          type: number
      required: [bookings, total]
...
    Notification:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        type:
          type: string
          enum: [booking, payment, document, support, system]
        createdAt:
          type: string
      required: [id, userId, title, message, isRead, type, createdAt]
    CreateNotificationBody:
      type: object
      properties:
        userId:
          type: string
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [booking, payment, document, support, system]
      required: [userId, title, message, type]
    NotificationListResponse:
      type: object
      properties:
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        unreadCount:
          type: number
      required: [notifications, unreadCount]
...